+++ src/ladb_opencutlist/ruby/constants.rb
@@ -2,6 +2,6 @@ module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
     EXTENSION_VERSION = '1.9.10-dev'.freeze
-    EXTENSION_BUILD = '202011220749'.freeze
+    EXTENSION_BUILD = '202011270838'.freeze
 
 end