+++ src/ladb_opencutlist/ruby/model/geom/section.rb
@@ -19,8 +19,8 @@ module Ladb::OpenCutList
 
     # -----
 
-    def to_max_precision_s
-      DimensionUtils.to_max_precision_s(@width) + ' x ' + DimensionUtils.to_max_precision_s(@height)
+    def to_ocl_precision_s
+      DimensionUtils.instance.to_ocl_precision_s(@width) + ' x ' + DimensionUtils.instance.to_ocl_precision_s(@height)
     end
 
     def to_s